+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+<<<<<<< HEAD
 ## 0.2.0 (2017-06-06)
 ### Added
 - Cassandra integration v0.1.0
@@ -15,3 +16,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1
 ### Added
 - Initial release
+=======
+## 1.0.0 (2017-07-27)
+
+Initial release
+
+### Added
+- Cassandra integration v1.0.0
+- NGINX integration v1.0.0
+- MySQL integration v1.0.0
+>>>>>>> upstream/master